refactor: set scope tree tabs (4) - Ref gestion-de-projet#2027
