+++ src/components/NewScopeTree/styles.js
@@ -2,9 +2,10 @@ import { makeStyles } from 'tss-react/mui'
 
 const useStyles = makeStyles()(() => ({
   tabs: {
-    marginTop: 72,
-    position: 'absolute',
     width: '100%'
+  },
+  searchBar: {
+    marginBottom: '0px'
   }
 }))
